Command Line Application/requirements.txt
tensorflow==2.1.0
tensorflow_hub==0.8.0
Pillow==7.1.1